requirements.txt
attrs==24.2.0
Automat==24.8.1
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.1
cssselect==1.2.0
defusedxml==0.7.1
filelock==3.16.1
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==3.1.0
greenlet==3.1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.9.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
lxml==5.3.0
MarkupSafe==3.0.1
nltk==3.9.1
numpy==2.1.2
packaging==24.1
parsel==1.9.1
Protego==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==24.2.1
queuelib==1.7.0
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
scikit-learn==1.5.2
scipy==1.14.1
Scrapy==2.11.2
service-identity==24.1.0
setuptools==75.1.0
soupsieve==2.6
SQLAlchemy==2.0.35
threadpoolctl==3.5.0
tldextract==5.1.2
tqdm==4.66.5
Twisted==24.7.0
typing_extensions==4.12.2
urllib3==2.2.3
w3lib==2.2.1
Werkzeug==3.0.4
zope.interface==7.0.3
PyPDF2==1.26.0
python-docx==0.8.11
Flask-Login==0.6.3
Flask-Migrate==3.1.0
Scrapy==2.11.2
python-dotenv==1.0.1